embed in float16

+++ mprompt/methods/m1_ngrams.py
@@ -38,29 +38,28 @@ def explain_ngrams(
 
     # compute scores and cache...
     # fmri should cache all preds together, since they are efficiently computed together
-    if use_cache:
-        if args.module_name == 'fmri':
-            cache_file = join(CACHE_DIR, 'cache_ngrams',
-                              f'{args.module_name}_{args.subject}.pkl')
-        if args.module_name == 'old_fmri':
-            cache_file = join(CACHE_DIR, 'cache_ngrams',
-                              f'{args.module_name}.pkl')
+    if args.module_name == 'fmri':
+        cache_file = join(CACHE_DIR, 'cache_ngrams',
+                            f'{args.module_name}_{args.subject}.pkl')
+    if args.module_name == 'old_fmri':
+        cache_file = join(CACHE_DIR, 'cache_ngrams',
+                            f'{args.module_name}.pkl')
+    else:
+        cache_file = join(CACHE_DIR, 'cache_ngrams',
+                            f'{args.module_name}_{args.module_num}.pkl')
+    if use_cache and os.path.exists(cache_file):
+        ngram_scores = pkl.load(open(cache_file, 'rb'))
+    else:
+        call_parameters = inspect.signature(mod.__call__).parameters.keys()
+        print('predicting all ngrams...')
+        if 'return_all' in call_parameters:
+            ngram_scores = mod(ngrams_list, return_all=True)
         else:
-            cache_file = join(CACHE_DIR, 'cache_ngrams',
-                              f'{args.module_name}_{args.module_num}.pkl')
-        if os.path.exists(cache_file):
-            ngram_scores = pkl.load(open(cache_file, 'rb'))
-        else:
-            call_parameters = inspect.signature(mod.__call__).parameters.keys()
-            print('predicting all ngrams...')
-            if 'return_all' in call_parameters:
-                ngram_scores = mod(ngrams_list, return_all=True)
-            else:
-                ngram_scores = mod(ngrams_list)
-
-            if use_cache:
-                os.makedirs(dirname(cache_file), exist_ok=True)
-                pkl.dump(ngram_scores, open(cache_file, 'wb'))
+            ngram_scores = mod(ngrams_list)
+            
+        if use_cache:
+            os.makedirs(dirname(cache_file), exist_ok=True)
+            pkl.dump(ngram_scores, open(cache_file, 'wb'))
 
     # multidimensional predictions
     if len(ngram_scores.shape) > 1 and ngram_scores.shape[1] > 1:
